Moved exceptions into own module and created common superclass.
This makes it easy to catch any "adaptor" error when running an import.
