+++ adaptor/model.py
@@ -5,7 +5,8 @@ import copy
 
 import csv
 from django.db.models.base import Model
-from fields import Field, ForeignKeyFieldError, IgnoredField, ComposedKeyField, XMLRootField, FieldValueMissing
+from adaptor.fields import Field, IgnoredField, ComposedKeyField, XMLRootField
+from adaptor.exceptions import ForeignKeyFieldError, FieldValueMissing
 
 
 class ImproperlyConfigured(Exception):
